pyproject.toml
[project]
name = "pakistan-lawer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "openai>=1.109.1",
    "pydantic>=2.11.9",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
